Backend/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.4
django-allauth==0.52.0
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==1.2.1
drf-yasg==1.21.4
ecdsa==0.18.0
filelock==3.8.2
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
lxml==4.9.2
MarkupSafe==2.1.2
myenv==201902.7
oauthlib==3.2.2
packaging==23.0
Pillow==9.3.0
pip==22.3.1
platformdirs==2.6.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-decouple==3.7
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
setuptools==65.5.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.17.1
virtualenvwrapper-win==1.2.7
wrapt==1.14.1